🐞 fix: Fixes for dotenv
You cannot inject variables using dotenv in a chrome extension.

Cleaning this up and general organization.
