+++ popup.html
@@ -8,11 +8,13 @@
 <body>
     <header>
       <h1>Job Application Tracker</h1>
-      <span id="sheet"></span>
-      <button id="linkSheet">
-        <!-- <img src="/assets/cogwheel32.png"/> -->
-        Save Google Sheet
-      </button>
+      <div id="sheetSection">
+        <span id="sheet"></span>
+        <button id="linkSheet">
+          <img src="/assets/cogwheel32.png"/>
+          <!-- Save Google Sheet -->
+        </button>
+      </div>
     </header>
     <hr />
     <div id="form">
